requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
cython==0.29.21
    # via n2
idna==2.10
    # via requests
joblib==0.17.0
    # via scikit-learn
more-itertools==8.6.0
    # via -r requirements.in
n2==0.1.7
    # via -r requirements.in
nlpaug==1.1.2
    # via -r requirements.in
numpy==1.19.4
    # via
    #   -r requirements.in
    #   pytorch-metric-learning
    #   scikit-learn
    #   scipy
    #   torch
    #   torchvision
ordered-set==4.0.2
    # via -r requirements.in
pillow==8.0.1
    # via torchvision
python-levenshtein==0.12.1
    # via -r requirements.in
pytorch-metric-learning==0.9.95
    # via -r requirements.in
regex==2020.11.13
    # via -r requirements.in
requests==2.25.1
    # via -r requirements.in
scikit-learn==0.23.2
    # via
    #   -r requirements.in
    #   pytorch-metric-learning
scipy==1.5.4
    # via scikit-learn
threadpoolctl==2.1.0
    # via scikit-learn
torch==1.7.1
    # via
    #   -r requirements.in
    #   pytorch-metric-learning
    #   torchvision
torchvision==0.8.2
    # via pytorch-metric-learning
tqdm==4.53.0
    # via
    #   -r requirements.in
    #   pytorch-metric-learning
typing-extensions==3.7.4.3
    # via torch
unidecode==1.1.2
    # via -r requirements.in
urllib3==1.26.2
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
